Functions moved to a new module
Added local Storage for wen the user reload the page
