+++ src/modules/checkbox-status.js
@@ -1,0 +1,12 @@
+// export const clearAllTasks = (list) => {
+//   const $containers = document.querySelectorAll('.task-container');
+//   list.forEach((task) => {
+//     if (task.completed) {
+//       $containers.forEach((element) => {
+//         if (task.index === parseInt(element.id, 10)) {
+//           element.remove();
+//         }
+//       });
+//     }
+//   });
+// };
